Refactor dbmate package to be importable as a package
