+++ pkg/driver/mysql/mysql_test.go
@@ -1,32 +1,17 @@
-package main
+package mysql
 
 import (
 	"database/sql"
 	"net/url"
-	"os"
 	"testing"
 
+	"github.com/flowhamster/dbmate/pkg/utils"
 	"github.com/stretchr/testify/require"
 )
 
-func postgresTestURL(t *testing.T) *url.URL {
-	str := os.Getenv("POSTGRES_PORT")
-	require.NotEmpty(t, str, "missing POSTGRES_PORT environment variable")
-
-	u, err := url.Parse(str)
-	require.Nil(t, err)
-
-	u.Scheme = "postgres"
-	u.User = url.User("postgres")
-	u.Path = "/dbmate"
-	u.RawQuery = "sslmode=disable"
-
-	return u
-}
-
-func prepTestPostgresDB(t *testing.T) *sql.DB {
-	drv := PostgresDriver{}
-	u := postgresTestURL(t)
+func prepTestMySQLDB(t *testing.T) *sql.DB {
+	drv := MySQLDriver{}
+	u := MySQLTestURL(t)
 
 	// drop any existing database
 	err := drv.DropDatabase(u)
@@ -37,15 +22,15 @@ func prepTestPostgresDB(t *testing.T) *sql.DB {
 	require.Nil(t, err)
 
 	// connect database
-	db, err := sql.Open("postgres", u.String())
+	db, err := drv.Open(u)
 	require.Nil(t, err)
 
 	return db
 }
 
-func TestPostgresCreateDropDatabase(t *testing.T) {
-	drv := PostgresDriver{}
-	u := postgresTestURL(t)
+func TestMySQLCreateDropDatabase(t *testing.T) {
+	drv := MySQLDriver{}
+	u := MySQLTestURL(t)
 
 	// drop any existing database
 	err := drv.DropDatabase(u)
@@ -57,9 +42,9 @@ func TestPostgresCreateDropDatabase(t *testing.T) {
 
 	// check that database exists and we can connect to it
 	func() {
-		db, err := sql.Open("postgres", u.String())
+		db, err := drv.Open(u)
 		require.Nil(t, err)
-		defer mustClose(db)
+		defer utils.MustClose(db)
 
 		err = db.Ping()
 		require.Nil(t, err)
@@ -71,19 +56,19 @@ func TestPostgresCreateDropDatabase(t *testing.T) {
 
 	// check that database no longer exists
 	func() {
-		db, err := sql.Open("postgres", u.String())
+		db, err := drv.Open(u)
 		require.Nil(t, err)
-		defer mustClose(db)
+		defer utils.MustClose(db)
 
 		err = db.Ping()
 		require.NotNil(t, err)
-		require.Equal(t, "pq: database \"dbmate\" does not exist", err.Error())
+		require.Regexp(t, "Unknown database 'dbmate'", err.Error())
 	}()
 }
 
-func TestPostgresDatabaseExists(t *testing.T) {
-	drv := PostgresDriver{}
-	u := postgresTestURL(t)
+func TestMySQLDatabaseExists(t *testing.T) {
+	drv := MySQLDriver{}
+	u := MySQLTestURL(t)
 
 	// drop any existing database
 	err := drv.DropDatabase(u)
@@ -104,25 +89,25 @@ func TestPostgresDatabaseExists(t *testing.T) {
 	require.Equal(t, true, exists)
 }
 
-func TestPostgresDatabaseExists_Error(t *testing.T) {
-	drv := PostgresDriver{}
-	u := postgresTestURL(t)
+func TestMySQLDatabaseExists_Error(t *testing.T) {
+	drv := MySQLDriver{}
+	u := MySQLTestURL(t)
 	u.User = url.User("invalid")
 
 	exists, err := drv.DatabaseExists(u)
-	require.Equal(t, "pq: role \"invalid\" does not exist", err.Error())
+	require.Regexp(t, "Access denied for user 'invalid'@", err.Error())
 	require.Equal(t, false, exists)
 }
 
-func TestPostgresCreateMigrationsTable(t *testing.T) {
-	drv := PostgresDriver{}
-	db := prepTestPostgresDB(t)
-	defer mustClose(db)
+func TestMySQLCreateMigrationsTable(t *testing.T) {
+	drv := MySQLDriver{}
+	db := prepTestMySQLDB(t)
+	defer utils.MustClose(db)
 
 	// migrations table should not exist
 	count := 0
 	err := db.QueryRow("select count(*) from schema_migrations").Scan(&count)
-	require.Equal(t, "pq: relation \"schema_migrations\" does not exist", err.Error())
+	require.Regexp(t, "Table 'dbmate.schema_migrations' doesn't exist", err.Error())
 
 	// create table
 	err = drv.CreateMigrationsTable(db)
@@ -137,10 +122,10 @@ func TestPostgresCreateMigrationsTable(t *testing.T) {
 	require.Nil(t, err)
 }
 
-func TestPostgresSelectMigrations(t *testing.T) {
-	drv := PostgresDriver{}
-	db := prepTestPostgresDB(t)
-	defer mustClose(db)
+func TestMySQLSelectMigrations(t *testing.T) {
+	drv := MySQLDriver{}
+	db := prepTestMySQLDB(t)
+	defer utils.MustClose(db)
 
 	err := drv.CreateMigrationsTable(db)
 	require.Nil(t, err)
@@ -163,10 +148,10 @@ func TestPostgresSelectMigrations(t *testing.T) {
 	require.Equal(t, false, migrations["abc2"])
 }
 
-func TestPostgresInsertMigration(t *testing.T) {
-	drv := PostgresDriver{}
-	db := prepTestPostgresDB(t)
-	defer mustClose(db)
+func TestMySQLInsertMigration(t *testing.T) {
+	drv := MySQLDriver{}
+	db := prepTestMySQLDB(t)
+	defer utils.MustClose(db)
 
 	err := drv.CreateMigrationsTable(db)
 	require.Nil(t, err)
@@ -186,10 +171,10 @@ func TestPostgresInsertMigration(t *testing.T) {
 	require.Equal(t, 1, count)
 }
 
-func TestPostgresDeleteMigration(t *testing.T) {
-	drv := PostgresDriver{}
-	db := prepTestPostgresDB(t)
-	defer mustClose(db)
+func TestMySQLDeleteMigration(t *testing.T) {
+	drv := MySQLDriver{}
+	db := prepTestMySQLDB(t)
+	defer utils.MustClose(db)
 
 	err := drv.CreateMigrationsTable(db)
 	require.Nil(t, err)
