+++ pkg/utils/utils.go
@@ -1,4 +1,4 @@
-package main
+package utils
 
 import (
 	"io"
@@ -6,7 +6,7 @@ import (
 )
 
 // databaseName returns the database name from a URL
-func databaseName(u *url.URL) string {
+func DatabaseName(u *url.URL) string {
 	name := u.Path
 	if len(name) > 0 && name[:1] == "/" {
 		name = name[1:len(name)]
@@ -15,7 +15,7 @@ func databaseName(u *url.URL) string {
 	return name
 }
 
-func mustClose(c io.Closer) {
+func MustClose(c io.Closer) {
 	if err := c.Close(); err != nil {
 		panic(err)
 	}
