+++ pkg/utils/utils_test.go
@@ -1,9 +1,10 @@
-package main
+package utils_test
 
 import (
 	"net/url"
 	"testing"
 
+	"github.com/flowhamster/dbmate/pkg/utils"
 	"github.com/stretchr/testify/require"
 )
 
@@ -11,7 +12,7 @@ func TestDatabaseName(t *testing.T) {
 	u, err := url.Parse("ignore://localhost/foo?query")
 	require.Nil(t, err)
 
-	name := databaseName(u)
+	name := utils.DatabaseName(u)
 	require.Equal(t, "foo", name)
 }
 
@@ -19,6 +20,6 @@ func TestDatabaseName_Empty(t *testing.T) {
 	u, err := url.Parse("ignore://localhost")
 	require.Nil(t, err)
 
-	name := databaseName(u)
+	name := utils.DatabaseName(u)
 	require.Equal(t, "", name)
 }
