+++ pkg/version/version.go
@@ -1,4 +1,4 @@
-package main
+package version
 
 // Version of dbmate
 const Version = "1.2.1"
